requirements.txt
Adafruit-Blinka==8.5.0
adafruit-circuitpython-busdevice==5.2.3
adafruit-circuitpython-neopixel==6.3.6
adafruit-circuitpython-pixelbuf==1.1.8
adafruit-circuitpython-requests==1.12.8
adafruit-circuitpython-typing==1.8.1
Adafruit-PlatformDetect==3.30.0
Adafruit-PureIO==1.1.9
black==22.8.0
click==8.1.3
keyboard==0.13.5
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
pyftdi==0.54.0
pyserial==3.5
pyusb==1.2.1
tomli==2.0.1
typing_extensions==4.3.0
